pollreactor: Move C pollreactor code from serialqueue.c to its own file
